TST add separate commont tests
